Removed Phase 12 Integration and Testing
I started to run the tasks in 12, but while implementing 12.2 it became
obvious that the planner's desire to create e2e tests didn't consider
the complexities of requiring real access to discord or a huge mock.

I may circle back at some point, but right now the complexity does not
justify the value.